# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lemon Music Open API",
    "version": "1.2.16.12",
    "description": "柠檬音乐自托管开放 API。路径同时挂载在 `/api` 与 `/api/v1`（推荐客户端使用 `/api/v1`）。\n\n鉴权：`Authorization: Bearer <sessionToken>`；音频流可用 `?ticket=` 短时效媒体票（也可兼容 `?token=` 会话 JWT）。\n\n曲目稳定 ID：`id` / `trackId`（UUID），路径变更后保持不变；仍返回 `filePath` 以兼容 Web。\n\n人读文档见 docs/open-api.md；交互文档 `GET /api/docs`。",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "推荐（开放客户端）"
    },
    {
      "url": "/api",
      "description": "兼容 Web"
    }
  ],
...
  "paths": {
...
    "/play/ticket": {
      "post": {
        "tags": [
          "play"
        ],
        "summary": "签发/续签媒体票",
        "description": "对本地曲目或流式相对 URL 签发短时效 ticket（默认 2h）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "401": {
            "description": "未登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string",
                    "description": "相对流式路径"
                  },
                  "trackId": {
                    "type": "string"
                  },
                  "localPath": {
                    "type": "string"
                  },
                  "ttlSec": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/play/url": {
      "post": {
        "tags": [
          "play"
        ],
        "summary": "解析播放地址",
        "description": "本地：传 trackId 或 localPath/filePath；在线：传 songId+source。返回的 url 带 ?ticket=",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "401": {
            "description": "未登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "trackId": {
                    "type": "string"
                  },
                  "localPath": {
                    "type": "string"
                  },
                  "filePath": {
                    "type": "string"
                  },
                  "songId": {
                    "type": "string"
                  },
                  "source": {
                    "type": "string"
                  },
                  "quality": {
                    "type": "string"
                  },
                  "smooth": {
                    "type": "boolean"
                  },
                  "refresh": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/play/local": {
      "get": {
        "tags": [
          "play"
        ],
        "summary": "本地音频流",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "description": "绝对路径",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ticket",
            "in": "query",
            "description": "媒体票",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "query",
            "description": "会话 JWT（兼容）",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "audio/* Range 流"
          }
        }
      }
    },
    "/play/local-ape": {
      "get": {
        "tags": [
          "play"
        ],
        "summary": "APE 转码流",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "description": "路径",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ticket",
            "in": "query",
            "description": "媒体票",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "401": {
            "description": "未登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/play/local-smooth": {
      "get": {
        "tags": [
          "play"
        ],
        "summary": "流畅 AAC 流",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "description": "路径",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ticket",
            "in": "query",
            "description": "媒体票",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "401": {
            "description": "未登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/play/smooth-warmup": {
      "post": {
        "tags": [
          "play"
        ],
        "summary": "预热下一首 AAC",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "401": {
            "description": "未登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/play/proxy": {
      "get": {
        "tags": [
          "play"
        ],
        "summary": "远程音频反代",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "description": "远程 URL",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "in": "query",
            "description": "平台",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ticket",
            "in": "query",
            "description": "媒体票",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "401": {
            "description": "未登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/play/lyric": {
      "post": {
        "tags": [
          "play"
        ],
        "summary": "获取歌词",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "401": {
            "description": "未登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/play/cover": {
      "post": {
        "tags": [
          "play"
        ],
        "summary": "获取封面元数据",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "401": {
            "description": "未登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/play/cover-img": {
      "get": {
        "tags": [
          "play"
        ],
        "summary": "封面图片",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "description": "封面 URL",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ticket",
            "in": "query",
            "description": "媒体票",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "image/*"
          }
        }
      }
    },
...
    "/tag/cover": {
      "get": {
        "tags": [
          "tag"
        ],
        "summary": "嵌入封面",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "description": "文件路径",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ticket",
            "in": "query",
            "description": "媒体票",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "image/*"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录返回的 session token"
      },
      "streamTicket": {
        "type": "apiKey",
        "in": "query",
        "name": "ticket",
        "description": "短时效媒体票（HMAC，默认 2h）"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
      "Ok": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "additionalProperties": true
      },
...
    }
  }
}